TopInfo had been bifurcated; manually merged files
